+++ src/common/interfaces/config/app-config.interface.ts
@@ -9,6 +9,7 @@ export interface IConfigAttributes {
 	dbPassword: string
 	dbRetryAttempts: number | string
 	jwtSecret: string
+	fileStorageLocation: string
 }
 
 export interface IConfig {