+++ src/modules/notes/notes.module.ts
@@ -1,9 +1,14 @@
+import { NotesService } from './notes.service'
 import { NotesController } from './notes.controller'
 import { Module } from '@nestjs/common'
+import { SequelizeModule } from '@nestjs/sequelize'
+import { Note } from './models/notes.model'
+import { NotesRepository } from './notes.repository'
 
 @Module({
-	imports: [],
+	imports: [SequelizeModule.forFeature([Note])],
 	controllers: [NotesController],
-	providers: []
+	providers: [NotesService, NotesRepository],
+	exports: [SequelizeModule]
 })
 export class NotesModule {}