+++ test/util/index.ts
@@ -1,5 +1,4 @@
 import axios, { AxiosRequestConfig } from 'axios'
-import crypto from 'crypto'
 
 export const TEST_AUTH_URL = 'http://localhost:5555/api/auth'
 export const TEST_USERNAME = 'test_sa@example.com'